[Driver][SYCL] Refactor traits macros settings to enable new model
The SYCL device traits settings were being performed within the old
model specific action builder class.  Move this to the SYCL toolchain
area to be accessible by both the new and old offloading model.  Also
hookup the new offloading model portion of this.
