+++ clang/lib/Driver/ToolChains/SYCL.cpp
@@ -17,6 +17,7 @@
 #include "llvm/Support/CommandLine.h"
 #include "llvm/Support/FileSystem.h"
 #include "llvm/Support/Path.h"
+#include "llvm/SYCLLowerIR/DeviceConfigFile.hpp"
 #include <algorithm>
 #include <sstream>
 
@@ -382,6 +383,100 @@ SYCL::getDeviceLibraries(const Compilation &C, const llvm::Triple &TargetTriple,
     addLibraries(SYCLNativeCpuDeviceLibs);
 
   return LibraryList;
+}
+
+/// Reads device config file to find information about the SYCL targets in
+/// `Targets`, and defines device traits macros accordingly.
+void SYCL::populateSYCLDeviceTraitsMacrosArgs(
+    Compilation &C, const llvm::opt::ArgList &Args,
+    const SmallVectorImpl<std::pair<const ToolChain *, StringRef>> &Targets) {
+  if (Targets.empty())
+    return;
+
+  const auto &TargetTable = DeviceConfigFile::TargetTable;
+  std::map<StringRef, unsigned int> AllDevicesHave;
+  std::map<StringRef, bool> AnyDeviceHas;
+  bool AnyDeviceHasAnyAspect = false;
+  unsigned int ValidTargets = 0;
+  for (const auto &[TC, BoundArch] : Targets) {
+    assert(TC && "Invalid SYCL Offload Toolchain");
+    // Try and find the device arch, if it's empty, try to search for either
+    // the whole Triple or just the 'ArchName' string.
+    auto TargetIt = TargetTable.end();
+    const llvm::Triple &TargetTriple = TC->getTriple();
+    const StringRef TargetArch{BoundArch};
+    if (!TargetArch.empty()) {
+      TargetIt = llvm::find_if(TargetTable, [&](const auto &Value) {
+        using namespace tools::SYCL;
+        StringRef Device{Value.first};
+        if (Device.consume_front(gen::AmdGPU))
+          return TargetArch == Device && TargetTriple.isAMDGCN();
+        if (Device.consume_front(gen::NvidiaGPU))
+          return TargetArch == Device && TargetTriple.isNVPTX();
+        if (Device.consume_front(gen::IntelGPU))
+          return TargetArch == Device && TargetTriple.isSPIRAOT();
+        return TargetArch == Device;
+      });
+    } else {
+      TargetIt = TargetTable.find(TargetTriple.str());
+      if (TargetIt == TargetTable.end())
+        TargetIt = TargetTable.find(TargetTriple.getArchName().str());
+    }
+
+    if (TargetIt != TargetTable.end()) {
+      const DeviceConfigFile::TargetInfo &Info = (*TargetIt).second;
+      ++ValidTargets;
+      const auto &AspectList = Info.aspects;
+      const auto &MaySupportOtherAspects = Info.maySupportOtherAspects;
+      if (!AnyDeviceHasAnyAspect)
+        AnyDeviceHasAnyAspect = MaySupportOtherAspects;
+      for (const auto &Aspect : AspectList) {
+        // If target has an entry in the config file, the set of aspects
+        // supported by all devices supporting the target is 'AspectList'.
+        // If there's no entry, such set is empty.
+        const auto &AspectIt = AllDevicesHave.find(Aspect);
+        if (AspectIt != AllDevicesHave.end())
+          ++AllDevicesHave[Aspect];
+        else
+          AllDevicesHave[Aspect] = 1;
+        // If target has an entry in the config file AND
+        // 'MaySupportOtherAspects' is false, the set of aspects supported
+        // by any device supporting the target is 'AspectList'. If there's
+        // no entry OR 'MaySupportOtherAspects' is true, such set contains
+        // all the aspects.
+        AnyDeviceHas[Aspect] = true;
+      }
+    }
+  }
+  // If there's no entry for the target in the device config file, the set
+  // of aspects supported by any device supporting the target contains all
+  // the aspects.
+  if (ValidTargets == 0)
+    AnyDeviceHasAnyAspect = true;
+
+  const Driver &D = C.getDriver();
+  if (AnyDeviceHasAnyAspect) {
+    // There exists some target that supports any given aspect.
+    constexpr static StringRef MacroAnyDeviceAnyAspect{
+        "-D__SYCL_ANY_DEVICE_HAS_ANY_ASPECT__=1"};
+    D.addSYCLDeviceTraitsMacroArg(Args, MacroAnyDeviceAnyAspect);
+  } else {
+    // Some of the aspects are not supported at all by any of the targets.
+    // Thus, we need to define individual macros for each supported aspect.
+    for (const auto &[TargetKey, SupportedTarget] : AnyDeviceHas) {
+      assert(SupportedTarget);
+      const SmallString<64> MacroAnyDevice{
+          {"-D__SYCL_ANY_DEVICE_HAS_", TargetKey, "__=1"}};
+      D.addSYCLDeviceTraitsMacroArg(Args, MacroAnyDevice);
+    }
+  }
+  for (const auto &[TargetKey, SupportedTargets] : AllDevicesHave) {
+    if (SupportedTargets != ValidTargets)
+      continue;
+    const SmallString<64> MacroAllDevices{
+        {"-D__SYCL_ALL_DEVICES_HAVE_", TargetKey, "__=1"}};
+    D.addSYCLDeviceTraitsMacroArg(Args, MacroAllDevices);
+  }
 }
 
 // The list should match pre-built SYCL device library files located in
